Work around api-platform failing to encode pdf error stack traces
In case an exception is thrown where there is PDF data in some variable on the
stack then api-platform tries to encode this into json/json-ld and fails because it's
binary data and not utf-8.

Instead of storing the error just catch the exception on layer above and rethrow as
a new exception.