Add advanced signature profile selection
We now have a list of profiles. Each profile is identified by a name
which needs to be passed to the API for selecting that profile.
Each profile contains a pdf-as key ID and profiel ID and each profile
is guarded by its own role (in this case a oauth scope translated to a role)