Rework the serialization group handling
We used to set one context at the "top" entity and then check for this in
all properties as well as embedded entities to decide whcih attributes should
be included and which shouldn't.

This is problematic because this adds a dependency from the embedded entity to the
embedding entity which adds a depencency between separate pacakges in the wrong
direction.

Instead we use "EntityName:output", "EntityName:input" in each entity and only
add them to the context in entities up in the dependency tree.