esign: Fix some signed PDFs having trailing newlines
Since the php internal soap support doesn't handle the pdf-as response format
we had to implement this ourselves. In case one part of the response did have
an empty trailing line it would leak out into the PDF data.

While this doesn't result in the signature to be invalid, some tools complain that not 100%
of the document is covered by the included signature.

Instead of doing the parsing ourselves we now create a fake mime header (which PHP strips) with some
guessing based on the input and pass the input to zbateson/mail-mime-parser which handles all those
corner cases for us and splits things up correctly.