Implement adding user defined text to signature blocks
Both for advanced and qualified sigantures the user can pass
a "user_text" param with the upooad request which contains a list
of description/value pairs. These then get added to the signature
block tables.

The config is extended to allow the integrator to specify which
table should be used for appending the content.