Add support for multiple qualified signature profiles
Similar to the advanced signature profiles. Pass the profile
name via the API to select the right config. This also includes
the role the users needs to have for it to succeed.

To make the API break transition smoother make the profile name optional and
default to "default" for now. Can be removed once all clients are ported.