Fix soap parsing in case of leading newlines in the response
Due to PHP throwing out the MIME header we reconstruct it by somewhat
parsing the content.

That code didn't handle in case the response started with a newline and
thus didn't find a the ID of the first part.

To fix this ignore any leading whitespace when searching for the boundary ID.

Thanks to Peter Guthy for debugging/providing a fix.

Fixes #2