Provide callback urls in the bundle directly
In the qualified signature case pdf-as requires some URLs which it will
redirect to after the signature process is complete. Up until now we
required users to host this separately but ideally we would cover
the whole process with this bundle and also document it here.

Now in case the the callback URLs are not given in the config we will
generate them pointing to the api server itself. Also the two options are
marked as deprecated.

One problem is that generating public URLs isn't always that easy,
like in the CLI case where we have no request. Also the host name can't
always be trusted by default (though we whitelist redirects in pdf-as,
so not an issue here).

We'll have to improve documentation/defaults/config so we can more reliably
generate public absolute URLs for all setups.