SignParameters: different workaround for hyphen elements
SignParameters has three child elements where the element name
contains a hyphen, which isn't a valid PHP name.

With PHP getting rid of dynamic properties we need to get rid of them.

Luckily php-soap will use __get() to look up the element names, so
we can fake them just for serializing. And since this type is only
used as a parameter and not a return type we only need to handle
serilization and not deserialization.

To make sure that php-soap picks it up enable tracinf for the soap
service during tests and test that the values are included in the
XML request.