verify: error out in case verification is disabled
Even if the user has the rights, in case the feature is not enabled
throw an error if verification is used.

Since this feature never was finished there is likely no having the role
anyway, but this makes it more consistent with the api-docs where those
endpoints are hidden if the feature is not enabled.