Implement new bundle authorization system
Instead of using Symfony roles for authz we now use our generic authz framework.

Instead of a symfony role per profile there is now a global authz config which
defines two roles, for signing and verifying in general. And one resource specific
role for signing with specific profiles.

In case the profile has a Symfony role set in the config it overrides all this,
to keep old configs working.

* A user can sign if the profile has a Symfony role set and the user has that
  Symfony role.
* A user can sign if no Symfony role is set, and they have ROLE_SIGNER and
  ROLE_PROFILE_SIGNER.
* A user can verify if they have ROLE_VERIFIER.