Add option to allow replacing the signature image with a user provided PNG
We want to start supporting "simple signatures" with pdf-as, as in an image
of a hand-drawn signature that is attached to a PDF, with some process metadata
like email address which received the PDF for signing. While an electronic
signature is not needed in this case, it can't hurt and pdf-as is good tool
to add images to PDFs at specific positions.

To make this work with pdf-as we need to set the siganture image in the profile
via SIG_LABEL. SIG_LABEL can either be a path to an image file, or a
base64 encoded image itself. This means we can use the config override feature to
allow overriding the image via the SOAP API.

This extends both the advanced and qualified methods to take a custom image
which it then will try to override.

We restrict this to PNG for starters, as pdf-as seems to convert JPEG to
lossless, which results in larter images and thus isn't useful. Also limit
the image file size just in case.

While both signing methods support this now we don't expose it via the HTTP API
yet, only via the CLI. In the future we might expose via the HTTP API and require
a profile to opt-in.

While at it this also adds a "--user-text" option to the CLI commands for adding
custom user text.