Add dbp:relay:esign:preview command for creating a PNG preview image
Using the "/visblock" of pdf-as we create a preview image of a specific
profile and store it as a png file.

Based on the resolution we pass in and the pixel dimensions we get
back we can also estimate the default signature block size in points.
(the size in points if no width override is passed to the signing API)